pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-policy-auditor"
version = "0.1.0"
description = "Audit AI documentation against NIST AI RMF, EU AI Act, and ISO 42001"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ai", "compliance", "grc", "nist", "eu-ai-act", "iso-42001", "trust-and-safety"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
semantic = [
    "sentence-transformers>=3.0.0",
]
pdf = [
    "pypdf>=4.0.0",
]
api = [
    "slowapi>=0.1.9",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "bandit>=1.7.9",
    "pip-audit>=2.7.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0",
    "pre-commit>=3.7.0",
]
all = [
    "ai-policy-auditor[semantic,pdf,api,dev]",
]

[project.scripts]
ai-policy-auditor = "src.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/ai-policy-auditor"
Repository = "https://github.com/your-org/ai-policy-auditor"
Issues = "https://github.com/your-org/ai-policy-auditor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml -v"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "S", "B"]
ignore = ["S101", "S311"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
